Move RssFeed and RssIgnores into separate files
