+++ include/rssparser.h
@@ -1,10 +1,10 @@
 #ifndef NEWSBOAT_RSSPARSER_H_
 #define NEWSBOAT_RSSPARSER_H_
 
+#include <memory>
 #include <string>
 
 #include "remoteapi.h"
-#include "rss.h"
 #include "rss/feed.h"
 
 namespace rsspp {
@@ -13,8 +13,12 @@ class Item;
 
 namespace newsboat {
 
-class ConfigContainer;
 class Cache;
+class ConfigContainer;
+class CurlHandle;
+class RssFeed;
+class RssIgnores;
+class RssItem;
 
 class RssParser {
 public:
