+++ src/cache.cpp
@@ -15,7 +15,6 @@
 #include "dbexception.h"
 #include "logger.h"
 #include "matcherexception.h"
-#include "rss.h"
 #include "rssfeed.h"
 #include "strprintf.h"
 #include "utils.h"
