+++ src/itemrenderer.cpp
@@ -4,7 +4,6 @@
 
 #include "configcontainer.h"
 #include "htmlrenderer.h"
-#include "rss.h"
 #include "rssfeed.h"
 #include "textformatter.h"
