+++ src/queuemanager.cpp
@@ -5,7 +5,6 @@
 
 #include "configpaths.h"
 #include "fmtstrformatter.h"
-#include "rss.h"
 #include "rssfeed.h"
 #include "stflpp.h"
 #include "utils.h"
