+++ src/rssparser.cpp
@@ -13,11 +13,11 @@
 #include "logger.h"
 #include "newsblurapi.h"
 #include "ocnewsapi.h"
-#include "rss.h"
 #include "rss/exception.h"
 #include "rss/parser.h"
 #include "rss/rssparser.h"
 #include "rssfeed.h"
+#include "rssignores.h"
 #include "strprintf.h"
 #include "ttrssapi.h"
 #include "utils.h"
