+++ test/rssfeed.cpp
@@ -1,268 +1,9 @@
-#include "rss.h"
+#include "rssfeed.h"
 
 #include "3rd-party/catch.hpp"
 #include "cache.h"
 #include "configcontainer.h"
-#include "rss/exception.h"
-#include "rss/parser.h"
-#include "rss/rssparser.h"
-#include "rssfeed.h"
 #include "rssparser.h"
-#include "test-helpers.h"
-
-TEST_CASE("Throws exception if file doesn't exist", "[rsspp::Parser]")
-{
-	using TestHelpers::ExceptionWithMsg;
-
-	rsspp::Parser p;
-
-	REQUIRE_THROWS_MATCHES(p.parse_file("data/non-existent.xml"),
-		rsspp::Exception,
-		ExceptionWithMsg<rsspp::Exception>("could not parse file"));
-}
-
-TEST_CASE("Throws exception if file can't be parsed", "[rsspp::Parser]")
-{
-	using TestHelpers::ExceptionWithMsg;
-
-	rsspp::Parser p;
-
-	REQUIRE_THROWS_MATCHES(p.parse_file("data/empty.xml"),
-		rsspp::Exception,
-		ExceptionWithMsg<rsspp::Exception>("could not parse file"));
-}
-
-TEST_CASE("Extracts data from RSS 0.91", "[rsspp::Parser]")
-{
-	rsspp::Parser p;
-	rsspp::Feed f;
-
-	REQUIRE_NOTHROW(f = p.parse_file("data/rss091_1.xml"));
-
-	REQUIRE(f.rss_version == rsspp::Feed::RSS_0_91);
-	REQUIRE(f.title == "Example Channel");
-	REQUIRE(f.description == "an example feed");
-	REQUIRE(f.link == "http://example.com/");
-	REQUIRE(f.language == "en");
-
-	REQUIRE(f.items.size() == 1u);
-	REQUIRE(f.items[0].title == "1 < 2");
-	REQUIRE(f.items[0].link == "http://example.com/1_less_than_2.html");
-	REQUIRE(f.items[0].description ==
-		"1 < 2, 3 < 4.\nIn HTML, <b> starts a bold phrase\nand you "
-		"start a link with <a href=\n");
-	REQUIRE(f.items[0].author == "");
-	REQUIRE(f.items[0].guid == "");
-}
-
-TEST_CASE("Extracts data from RSS 0.92", "[rsspp::Parser]")
-{
-	rsspp::Parser p;
-	rsspp::Feed f;
-
-	REQUIRE_NOTHROW(f = p.parse_file("data/rss092_1.xml"));
-
-	REQUIRE(f.rss_version == rsspp::Feed::RSS_0_92);
-	REQUIRE(f.title == "Example Channel");
-	REQUIRE(f.description == "an example feed");
-	REQUIRE(f.link == "http://example.com/");
-	REQUIRE(f.language == "en");
-
-	REQUIRE(f.items.size() == 3u);
-
-	REQUIRE(f.items[0].title == "1 < 2");
-	REQUIRE(f.items[0].link == "http://example.com/1_less_than_2.html");
-	REQUIRE(f.items[0].base == "http://example.com/feed/rss_testing.html");
-
-	REQUIRE(f.items[1].title == "A second item");
-	REQUIRE(f.items[1].link == "http://example.com/a_second_item.html");
-	REQUIRE(f.items[1].description == "no description");
-	REQUIRE(f.items[1].author == "");
-	REQUIRE(f.items[1].guid == "");
-	REQUIRE(f.items[1].base == "http://example.com/item/rss_testing.html");
-
-	REQUIRE(f.items[2].title == "A third item");
-	REQUIRE(f.items[2].link == "http://example.com/a_third_item.html");
-	REQUIRE(f.items[2].description == "no description");
-	REQUIRE(f.items[2].base == "http://example.com/desc/rss_testing.html");
-}
-
-TEST_CASE("Extracts data fro RSS 2.0", "[rsspp::Parser]")
-{
-	rsspp::Parser p;
-	rsspp::Feed f;
-
-	REQUIRE_NOTHROW(f = p.parse_file("data/rss20_1.xml"));
-
-	REQUIRE(f.title == "my weblog");
-	REQUIRE(f.link == "http://example.com/blog/");
-	REQUIRE(f.description == "my description");
-
-	REQUIRE(f.items.size() == 1u);
-
-	REQUIRE(f.items[0].title == "this is an item");
-	REQUIRE(f.items[0].link ==
-		"http://example.com/blog/this_is_an_item.html");
-	REQUIRE(f.items[0].author == "Andreas Krennmair");
-	REQUIRE(f.items[0].author_email == "blog@synflood.at");
-	REQUIRE(f.items[0].content_encoded == "oh well, this is the content.");
-	REQUIRE(f.items[0].pubDate == "Fri, 12 Dec 2008 02:36:10 +0100");
-	REQUIRE(f.items[0].guid ==
-		"http://example.com/blog/this_is_an_item.html");
-	REQUIRE_FALSE(f.items[0].guid_isPermaLink);
-}
-
-TEST_CASE("Extracts data from RSS 1.0", "[rsspp::Parser]")
-{
-	rsspp::Parser p;
-	rsspp::Feed f;
-
-	REQUIRE_NOTHROW(f = p.parse_file("data/rss10_1.xml"));
-
-	REQUIRE(f.rss_version == rsspp::Feed::RSS_1_0);
-
-	REQUIRE(f.title == "Example Dot Org");
-	REQUIRE(f.link == "http://www.example.org");
-	REQUIRE(f.description == "the Example Organization web site");
-
-	REQUIRE(f.items.size() == 1u);
-
-	REQUIRE(f.items[0].title == "New Status Updates");
-	REQUIRE(f.items[0].link == "http://www.example.org/status/foo");
-	REQUIRE(f.items[0].guid == "http://www.example.org/status/");
-	REQUIRE(f.items[0].description == "News about the Example project");
-	REQUIRE(f.items[0].pubDate == "Tue, 30 Dec 2008 07:20:00 +0000");
-}
-
-TEST_CASE("Extracts data from Atom 1.0", "[rsspp::Parser]")
-{
-	rsspp::Parser p;
-	rsspp::Feed f;
-
-	REQUIRE_NOTHROW(f = p.parse_file("data/atom10_1.xml"));
-
-	REQUIRE(f.rss_version == rsspp::Feed::ATOM_1_0);
-
-	REQUIRE(f.title == "test atom");
-	REQUIRE(f.title_type == "text");
-	REQUIRE(f.description == "atom description!");
-	REQUIRE(f.pubDate == "Tue, 30 Dec 2008 18:26:15 +0000");
-	REQUIRE(f.link == "http://example.com/");
-
-	REQUIRE(f.items.size() == 3u);
-	REQUIRE(f.items[0].title == "A gentle introduction to Atom testing");
-	REQUIRE(f.items[0].title_type == "html");
-	REQUIRE(f.items[0].link == "http://example.com/atom_testing.html");
-	REQUIRE(f.items[0].guid == "tag:example.com,2008-12-30:/atom_testing");
-	REQUIRE(f.items[0].description == "some content");
-	REQUIRE(f.items[0].base == "http://example.com/feed/atom_testing.html");
-
-	REQUIRE(f.items[1].title == "A missing rel attribute");
-	REQUIRE(f.items[1].title_type == "html");
-	REQUIRE(f.items[1].link == "http://example.com/atom_testing.html");
-	REQUIRE(f.items[1].guid == "tag:example.com,2008-12-30:/atom_testing1");
-	REQUIRE(f.items[1].description == "some content");
-	REQUIRE(f.items[1].base ==
-		"http://example.com/entry/atom_testing.html");
-
-	REQUIRE(f.items[2].title == "alternate link isn't first");
-	REQUIRE(f.items[2].title_type == "html");
-	REQUIRE(f.items[2].link == "http://example.com/atom_testing.html");
-	REQUIRE(f.items[2].guid == "tag:example.com,2008-12-30:/atom_testing2");
-	REQUIRE(f.items[2].description == "some content");
-	REQUIRE(f.items[2].base ==
-		"http://example.com/content/atom_testing.html");
-}
-
-TEST_CASE("W3CDTF parser extracts date and time from any valid string",
-	"[rsspp::RssParser]")
-{
-	SECTION("year only")
-	{
-		REQUIRE(rsspp::RssParser::__w3cdtf_to_rfc822("2008") ==
-			"Tue, 01 Jan 2008 00:00:00 +0000");
-	}
-
-	SECTION("year-month only")
-	{
-		REQUIRE(rsspp::RssParser::__w3cdtf_to_rfc822("2008-12") ==
-			"Mon, 01 Dec 2008 00:00:00 +0000");
-	}
-
-	SECTION("year-month-day only")
-	{
-		REQUIRE(rsspp::RssParser::__w3cdtf_to_rfc822("2008-12-30") ==
-			"Tue, 30 Dec 2008 00:00:00 +0000");
-	}
-
-	SECTION("date and time with Z timezone")
-	{
-		REQUIRE(rsspp::RssParser::__w3cdtf_to_rfc822(
-				"2008-12-30T13:03:15Z") ==
-			"Tue, 30 Dec 2008 13:03:15 +0000");
-	}
-
-	SECTION("date and time with -08:00 timezone")
-	{
-		REQUIRE(rsspp::RssParser::__w3cdtf_to_rfc822(
-				"2008-12-30T10:03:15-08:00") ==
-			"Tue, 30 Dec 2008 18:03:15 +0000");
-	}
-}
-
-TEST_CASE("W3CDTF parser returns empty string on invalid input",
-	"[rsspp::RssParser]")
-{
-	REQUIRE(rsspp::RssParser::__w3cdtf_to_rfc822("foobar") == "");
-	REQUIRE(rsspp::RssParser::__w3cdtf_to_rfc822("-3") == "");
-	REQUIRE(rsspp::RssParser::__w3cdtf_to_rfc822("") == "");
-}
-
-TEST_CASE(
-	"W3C DTF to RFC 822 conversion does not take into account the local "
-	"timezone (#369)",
-	"[rsspp::RssParser]")
-{
-	auto input = "2008-12-30T10:03:15-08:00";
-	auto expected = "Tue, 30 Dec 2008 18:03:15 +0000";
-
-	TestHelpers::EnvVar tzEnv("TZ");
-	tzEnv.on_change([](){ ::tzset(); });
-
-	// US/Pacific and Australia/Sydney have pretty much opposite DST
-	// schedules, so for any given moment in time one of the following two
-	// sections will be observing DST while other won't
-	SECTION("Timezone Pacific")
-	{
-		tzEnv.set("US/Pacific");
-		REQUIRE(rsspp::RssParser::__w3cdtf_to_rfc822(input) ==
-			expected);
-	}
-
-	SECTION("Timezone Australia")
-	{
-		tzEnv.set("Australia/Sydney");
-		REQUIRE(rsspp::RssParser::__w3cdtf_to_rfc822(input) ==
-			expected);
-	}
-
-	// During October, both US/Pacific and Australia/Sydney are observing
-	// DST. Arizona and UTC *never* observe it, though, so the following two
-	// tests will cover October
-	SECTION("Timezone Arizona")
-	{
-		tzEnv.set("US/Arizona");
-		REQUIRE(rsspp::RssParser::__w3cdtf_to_rfc822(input) ==
-			expected);
-	}
-
-	SECTION("Timezone UTC")
-	{
-		tzEnv.set("UTC");
-		REQUIRE(rsspp::RssParser::__w3cdtf_to_rfc822(input) ==
-			expected);
-	}
-}
 
 namespace newsboat {
 
@@ -283,28 +24,6 @@ TEST_CASE("set_rssurl checks if query feed has a valid query", "[rss]")
 		REQUIRE_NOTHROW(f.set_rssurl("query:a title:unread = \"yes\""));
 		REQUIRE_NOTHROW(f.set_rssurl(
 			"query:Title:unread = \"yes\" and age between 0:7"));
-	}
-}
-
-TEST_CASE("RssItem::sort_flags() cleans up flags", "[rss]")
-{
-	ConfigContainer cfg;
-	Cache rsscache(":memory:", &cfg);
-	RssItem item(&rsscache);
-
-	SECTION("Repeated letters do not erase other letters")
-	{
-		std::string inputflags = "Abcdecf";
-		std::string result = "Abcdef";
-		item.set_flags(inputflags);
-		REQUIRE(result == item.flags());
-	}
-
-	SECTION("Non alpha characters in input flags are ignored")
-	{
-		std::string inputflags = "Abcd";
-		item.set_flags(inputflags + "1234568790^\"#'é(£");
-		REQUIRE(inputflags == item.flags());
 	}
 }
 
@@ -642,42 +361,8 @@ TEST_CASE(
 	REQUIRE(f.total_item_count() == 3);
 }
 
-TEST_CASE(
-	"RssIgnores::matches_lastmodified() returns true if given url "
-	"has to always be downloaded",
-	"[rss]")
-{
-	RssIgnores ignores;
-	ignores.handle_action("always-download",
-		{"http://newsboat.org",
-			"www.cool-website.com",
-			"www.example.com"});
-
-	REQUIRE(ignores.matches_lastmodified("www.example.com"));
-	REQUIRE(ignores.matches_lastmodified("http://newsboat.org"));
-	REQUIRE(ignores.matches_lastmodified("www.cool-website.com"));
-	REQUIRE_FALSE(ignores.matches_lastmodified("www.smth.com"));
-}
-
-TEST_CASE(
-	"RssIgnores::matches_resetunread() returns true if given url "
-	"will always have its unread flag reset on update",
-	"[rss]")
-{
-	RssIgnores ignores;
-	ignores.handle_action("reset-unread-on-update",
-		{"http://newsboat.org",
-			"www.cool-website.com",
-			"www.example.com"});
-
-	REQUIRE(ignores.matches_resetunread("www.example.com"));
-	REQUIRE(ignores.matches_resetunread("http://newsboat.org"));
-	REQUIRE(ignores.matches_resetunread("www.cool-website.com"));
-	REQUIRE_FALSE(ignores.matches_resetunread("www.smth.com"));
-}
-
 TEST_CASE("If item's <title> is empty, try to deduce it from the URL",
-	"[rss::RssParser]")
+	"[rss]")
 {
 	ConfigContainer cfg;
 	Cache rsscache(":memory:", &cfg);
